+++ src/Compute/Compute/Extension/Diagnostics/SetAzureRmVMDiagnosticsExtension.cs
@@ -176,7 +176,7 @@ namespace Microsoft.Azure.Commands.Compute
             }
         }
 
-        [Parameter(Mandatory = false, HelpMessage = "Starts the operation and returns immediately, before the operation is completed. In order to determine if the operation has sucessufuly been completed, use some other mechanism.")]
+        [Parameter(Mandatory = false, HelpMessage = "Starts the operation and returns immediately, before the operation is completed. In order to determine if the operation has successfully been completed, use some other mechanism.")]
         public SwitchParameter NoWait { get; set; }
 
         private Hashtable PublicConfiguration