+++ src/Monitor/Monitor/ChangeLog.md
@@ -19,6 +19,7 @@
 -->
 ## Upcoming Release
 * Fixed miscellaneous typos across module
+* Fixed incorrect parameter name in help documentation
 
 ## Version 1.2.1
 * Fixed incorrect parameter names in help examples