+++ src/Monitor/Monitor/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Fixed miscellaneous typos across module
+* Fixed incorrect parameter name in help documentation
 
 ## Version 1.2.1
 * Fixed incorrect parameter names in help examples